requirements.txt
flask
uvicorn
pytest
re
python-dotenv